refactored google api interaction, moved text recognition to converter
